+++ src/app/docs/page/page.component.html
@@ -3,8 +3,8 @@
     <div *ngIf="selectedDoc">
     <!-- main content area -->
     <div class="col-md-9">
-      <div *ngIf="slug == 'blog'">
-        <h1>Twitter Time line</h1>
+      <div *ngIf="slug === 'blog'">
+        <h1>Twitter Timeline</h1>
         <a class="twitter-timeline"
            data-width="650"
            data-height="500"
