+++ src/app/home/home.component.html
@@ -54,7 +54,7 @@
     </div>
   </div>
 </div>
-<div class="container" style="padding-top:30px;" #mainScreen>
+<div class="container mainScreen" #mainScreen>
   <div class="row">
     <div class="col-md-9">
         <tabset class="homeComponent" type="tabs">
@@ -76,7 +76,7 @@
           </tab>
         </tabset>
     </div>
-    <div class="col-md-3" style="border: 1px solid #CFD1D7; border-radius: 15px">
+    <div class="col-md-3 twitter">
         <a class="twitter-timeline"
            data-theme="light"
            data-height="875"
