+++ src/app/home/home.component.scss
@@ -1,3 +1,10 @@
+.mainScreen {
+  padding-top:30px;
+}
+.twitter {
+  border: 1px solid #CFD1D7;
+  border-radius: 15px;
+}
 /* Video Button */
 .youtube, .youtube:active, .youtube:focus {
   float: none;
